session: extract models from utility functions
